// src/Types/CustomMaterial.ts
export type DbMaterial = {
    alphaTest: number;
    alphaToCoverage: boolean;
    blendDst: BlendingDstFactor;
    blendDstAlpha: number | null;
    blendEquation: BlendingEquation;
    blendEquationAlpha: number | null;
    blending: string|null;
    blendSrc: BlendingSrcFactor | BlendingDstFactor; 
    blendSrcAlpha: number | null;
    clipIntersection: boolean;
    clippingPlanes: any;
    clipShadows: boolean;
    colorWrite: boolean;
    defines: string | undefined;
    depthFunc: DepthModes;
    depthTest: boolean;
    depthWrite: boolean;
    fog: boolean;
    format: number;
    stencilWrite: boolean;
    stencilFunc: StencilFunc;
    stencilRef: number;  
    stencilWriteMask: number;   
    stencilFuncMask: number;   
    stencilFail: StencilOp;  
    stencilZFail: StencilOp;  
    stencilZPass: StencilOp;
    name: string;
    needsUpdate: boolean;
    opacity: number;
    polygonOffset: boolean;    
    polygonOffsetFactor: number;   
    polygonOffsetUnits: number;    
    precision: 'highp' | 'mediump' | 'lowp' | null;    
    premultipliedAlpha: boolean;    
    dithering: boolean;  
    side: Side;   
    shadowSide: Side | null;   
    toneMapped: boolean;
    transparent: boolean;    
    objectType: string;    
    id: string;   
    vertexColors: boolean;    
    visible: boolean;   
    userData: string;
    version: number;
    color?: string;
    map?: Texture | null | undefined;
    lightMap?: Texture | null;
    lightMapIntensity?: number | undefined;
    aoMap?: Texture | null | undefined;
    aoMapIntensity?: number | undefined;
    specularMap?: Texture | null | undefined;
    alphaMap?: Texture | null | undefined;
    envMap?: Texture | null | undefined;
    combine?: number | undefined;
    reflectivity?: number | undefined;
    refractionRatio?: number | undefined;
    wireframe?: boolean | undefined;
    wireframeLinewidth?: number | undefined;
    wireframeLinecap?: string | undefined;
    wireframeLinejoin?: string | undefined;
    rotation?: number | undefined;
    sizeAttenuation?: boolean | undefined;
    uniforms?: { [uniform: string]: IUniform } | undefined;
    vertexShader?: string | undefined;
    fragmentShader?: string | undefined;
    linewidth?: number | undefined;
    lights?: boolean | undefined;
    clipping?: boolean | undefined;
    extensions?:
        | {
                derivatives?: boolean | undefined;
                fragDepth?: boolean | undefined;
                drawBuffers?: boolean | undefined;
                shaderTextureLOD?: boolean | undefined;
            }
        | undefined;
    glslVersion?: GLSLVersion | undefined;
    size?: number | undefined;
    gradientMap?: Texture | null | undefined;
    emissive?: ColorRepresentation | undefined;
    emissiveIntensity?: number | undefined;
    emissiveMap?: Texture | null | undefined;
    bumpMap?: Texture | null | undefined;
    bumpScale?: number | undefined;
    normalMap?: Texture | null | undefined;
    normalMapType?: NormalMapTypes | undefined;
    normalScale?: Vector2 | undefined;
    displacementMap?: Texture | null | undefined;
    displacementScale?: number | undefined;
    displacementBias?: number | undefined;
    roughness?: number | undefined;
    metalness?: number | undefined;
    roughnessMap?: Texture | null | undefined;
    metalnessMap?: Texture | null | undefined;
    envMapIntensity?: number | undefined;
    flatShading?: boolean | undefined;
    specular?: ColorRepresentation | undefined;
    shininess?: number | undefined;
    clearcoat?: number | undefined;
    clearcoatMap?: Texture | null | undefined;
    clearcoatRoughness?: number | undefined;
    clearcoatRoughnessMap?: Texture | null | undefined;
    clearcoatNormalScale?: Vector2 | undefined;
    clearcoatNormalMap?: Texture | null | undefined;
    ior?: number | undefined;
    sheen?: number | undefined;
    sheenRoughness?: number | undefined;
    transmission?: number | undefined;
    transmissionMap?: Texture | null | undefined;
    attenuationDistance?: number | undefined;
    specularIntensity?: number | undefined;
    specularIntensityMap?: Texture | null | undefined;
    specularColorMap?: Texture | null | undefined;
    matcap?: Texture | null | undefined;
    farDistance?: number | undefined;
    nearDistance?: number | undefined;
    referencePosition?: Vector3 | undefined;
    depthPacking?: DepthPackingStrategies | undefined;
    scale?: number | undefined;
    dashSize?: number | undefined;
    gapSize?: number | undefined;
    linecap?: string | undefined;
    linejoin?: string | undefined;

}

export interface CombinedMaterial extends Material {
    color?: Color | undefined;
    map?: Texture | null | undefined;
    lightMap?: Texture | null;
    lightMapIntensity?: number | undefined;
    aoMap?: Texture | null | undefined;
    aoMapIntensity?: number | undefined;
    specularMap?: Texture | null | undefined;
    alphaMap?: Texture | null | undefined;
    envMap?: Texture | null | undefined;
    combine?: Combine | undefined;
    reflectivity?: number | undefined;
    refractionRatio?: number | undefined;
    wireframe?: boolean | undefined;
    wireframeLinewidth?: number | undefined;
    wireframeLinecap?: string | undefined;
    wireframeLinejoin?: string | undefined;
    rotation?: number | undefined;
    sizeAttenuation?: boolean | undefined;
    uniforms?: { [uniform: string]: IUniform } | undefined;
    vertexShader?: string | undefined;
    fragmentShader?: string | undefined;
    linewidth?: number | undefined;
    lights?: boolean | undefined;
    clipping?: boolean | undefined;
    extensions?:
        | {
              derivatives?: boolean | undefined;
              fragDepth?: boolean | undefined;
              drawBuffers?: boolean | undefined;
              shaderTextureLOD?: boolean | undefined;
          }
        | undefined;
    glslVersion?: GLSLVersion | undefined;
    size?: number | undefined;
    gradientMap?: Texture | null | undefined;
    emissive?: ColorRepresentation | undefined;
    emissiveIntensity?: number | undefined;
    emissiveMap?: Texture | null | undefined;
    bumpMap?: Texture | null | undefined;
    bumpScale?: number | undefined;
    normalMap?: Texture | null | undefined;
    normalMapType?: NormalMapTypes | undefined;
    normalScale?: Vector2 | undefined;
    displacementMap?: Texture | null | undefined;
    displacementScale?: number | undefined;
    displacementBias?: number | undefined;
    roughness?: number | undefined;
    metalness?: number | undefined;
    roughnessMap?: Texture | null | undefined;
    metalnessMap?: Texture | null | undefined;
    envMapIntensity?: number | undefined;
    flatShading?: boolean | undefined;
    specular?: ColorRepresentation | undefined;
    shininess?: number | undefined;
    clearcoat?: number | undefined;
    clearcoatMap?: Texture | null | undefined;
    clearcoatRoughness?: number | undefined;
    clearcoatRoughnessMap?: Texture | null | undefined;
    clearcoatNormalScale?: Vector2 | undefined;
    clearcoatNormalMap?: Texture | null | undefined;
    ior?: number | undefined;
    sheen?: number | undefined;
    sheenRoughness?: number | undefined;
    transmission?: number | undefined;
    transmissionMap?: Texture | null | undefined;
    attenuationDistance?: number | undefined;
    specularIntensity?: number | undefined;
    specularIntensityMap?: Texture | null | undefined;
    specularColorMap?: Texture | null | undefined;
    matcap?: Texture | null | undefined;
    farDistance?: number | undefined;
    nearDistance?: number | undefined;
    referencePosition?: Vector3 | undefined;
    depthPacking?: DepthPackingStrategies | undefined;
    scale?: number | undefined;
    dashSize?: number | undefined;
    gapSize?: number | undefined;
    linecap?: string | undefined;
    linejoin?: string | undefined;
}